model-training/requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
opencv-python==4.8.1.78
tensorflowjs==4.14.0
